[005]-File_upload

+++ src/main/java/B4F2/PetStagram/feed/service/FeedService.java
@@ -37,8 +37,12 @@ public class FeedService {
 
     public boolean deleteFeed(Long feedId, String name) {
 
-        feedRepository.findById(feedId)
+        FeedEntity feed = feedRepository.findById(feedId)
                 .orElseThrow(() -> new CustomException(ErrorCode.NOT_FOUND_BOARD));
+
+        if (!name.equals(feed.getUserId())) {
+            throw new CustomException(ErrorCode.UN_MATCH_ID);
+        }
 
         feedRepository.deleteById(feedId);
